Moving auth_tests to infra
