+++ tests/test_file.py
@@ -22,10 +22,10 @@ from dss.config import BucketConfig, Config, override_bucket_config, Replica
 from dss.storage.hcablobstore import compose_blob_key
 from dss.util import UrlBuilder
 from dss.util.version import datetime_to_version_format
-from tests.auth_tests import TestAuthMixin
 from tests import eventually, get_auth_header
 from tests.fixtures.cloud_uploader import GSUploader, S3Uploader, Uploader
-from tests.infra import DSSAssertMixin, DSSUploadMixin, ExpectedErrorFields, get_env, generate_test_key, testmode
+from tests.infra import DSSAssertMixin, DSSUploadMixin, ExpectedErrorFields, get_env, generate_test_key, testmode, \
+    TestAuthMixin
 from tests.infra.server import ThreadedLocalServer
